requirements.txt
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
pdfminer.six==20200726
pycparser==2.20
PyQt5==5.15.1
PyQt5-sip==12.8.1
six==1.15.0
sortedcontainers==2.2.2
